refactor(cascade-prelude): better api for text
